Fix parsing coverage report with modern coverage
At some point, coverage has started outputting the TOTAL line even when
you are only looking at coverage of a single file. This confuses the
coverage report parsing code, which only looks at the last line.

Change it to consider any line in the output, and succeed if it finds a
line matching the file name it is looking for.